chore(Refactoring): Resolved Insufficent modularization smell in TypeMetrics class
