+++ src/Designite/SourceModel/SM_Package.java
@@ -7,6 +7,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import Designite.metrics.TypeMetricsExtractor;
 import org.eclipse.jdt.core.dom.CompilationUnit;
 
 import Designite.InputArgs;
@@ -117,8 +118,7 @@ public class SM_Package extends SM_SourceItem implements Parsable{
 	public void extractTypeMetrics() {
 		for (SM_Type type : typeList) {
 			type.extractMethodMetrics();
-			TypeMetrics metrics = new TypeMetrics(type);
-			metrics.extractMetrics();
+			TypeMetrics metrics = new TypeMetricsExtractor(type).extractMetrics();
 			metricsMapping.put(type, metrics);
 			exportMetricsToCSV(metrics, type.getName());
 			updateDependencyGraph(type);
